// src/types/record.d.ts
export interface Record {
  id: string;
  studentId: string;
  studentName: string;
  points: number;  // تم إضافة حقل النقاط
  reason: string;
  teacher: string;
  dateTime: string;
  date: string;
  phoneNumber: string;
  teacherName: string;
  totalPoints: number;
  level: string;
  badge: string;
}